#117 - reducing SystemEliteId intermodule communications class amount
