+++ trade-module/src/test/java/io/edpn/backend/trade/application/service/CleanUpObsoleteSystemEliteIdRequestsUseCaseTest.java
@@ -1,19 +1,28 @@
 package io.edpn.backend.trade.application.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.SystemDataRequest;
 import io.edpn.backend.trade.application.domain.System;
 import io.edpn.backend.trade.application.domain.filter.FindSystemFilter;
+import io.edpn.backend.trade.application.dto.web.object.mapper.MessageMapper;
+import io.edpn.backend.trade.application.port.outgoing.kafka.SendKafkaMessagePort;
+import io.edpn.backend.trade.application.port.outgoing.system.LoadOrCreateSystemByNamePort;
 import io.edpn.backend.trade.application.port.outgoing.system.LoadSystemsByFilterPort;
+import io.edpn.backend.trade.application.port.outgoing.system.UpdateSystemPort;
 import io.edpn.backend.trade.application.port.outgoing.systemeliteidrequest.CleanUpObsoleteSystemEliteIdRequestsUseCase;
+import io.edpn.backend.trade.application.port.outgoing.systemeliteidrequest.CreateSystemEliteIdRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.systemeliteidrequest.DeleteSystemEliteIdRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.systemeliteidrequest.ExistsSystemEliteIdRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.systemeliteidrequest.LoadAllSystemEliteIdRequestsPort;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Executor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.retry.support.RetryTemplate;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -28,18 +37,47 @@ public class CleanUpObsoleteSystemEliteIdRequestsUseCaseTest {
 
     @Mock
     private LoadSystemsByFilterPort loadSystemsByFilterPort;
-
     @Mock
-    private LoadAllSystemEliteIdRequestsPort allSystemEliteIdRequestsPort;
-
+    private LoadAllSystemEliteIdRequestsPort loadAllSystemEliteIdRequestsPort;
+    @Mock
+    private LoadOrCreateSystemByNamePort loadOrCreateSystemByNamePort;
+    @Mock
+    private ExistsSystemEliteIdRequestPort existsSystemEliteIdRequestPort;
+    @Mock
+    private CreateSystemEliteIdRequestPort createSystemEliteIdRequestPort;
     @Mock
     private DeleteSystemEliteIdRequestPort deleteSystemEliteIdRequestPort;
+    @Mock
+    private UpdateSystemPort updateSystemPort;
+    @Mock
+    private SendKafkaMessagePort sendKafkaMessagePort;
+    @Mock
+    private RetryTemplate retryTemplate;
+    @Mock
+    private Executor executor;
+    @Mock
+    private ObjectMapper objectMapper;
+    @Mock
+    private MessageMapper messageMapper;
 
     private CleanUpObsoleteSystemEliteIdRequestsUseCase underTest;
 
     @BeforeEach
     public void setUp() {
-        underTest = new CleanUpObsoleteSystemEliteIdRequestsService(loadSystemsByFilterPort, allSystemEliteIdRequestsPort, deleteSystemEliteIdRequestPort);
+        underTest = new SystemEliteIdInterModuleCommunicationService(
+                loadSystemsByFilterPort,
+                loadAllSystemEliteIdRequestsPort,
+                loadOrCreateSystemByNamePort,
+                existsSystemEliteIdRequestPort,
+                createSystemEliteIdRequestPort,
+                deleteSystemEliteIdRequestPort,
+                updateSystemPort,
+                sendKafkaMessagePort,
+                retryTemplate,
+                executor,
+                objectMapper,
+                messageMapper
+        );
     }
 
     @Test
@@ -48,12 +86,12 @@ public class CleanUpObsoleteSystemEliteIdRequestsUseCaseTest {
                 .hasEliteId(false)
                 .build();
 
-        assertThat(CleanUpObsoleteSystemEliteIdRequestsService.FIND_SYSTEM_FILTER, equalTo(findSystemFilter));
+        assertThat(SystemEliteIdInterModuleCommunicationService.FIND_SYSTEM_FILTER, equalTo(findSystemFilter));
     }
 
     @Test
     public void testCleanUpObsolete_NoOpenRequests() {
-        when(allSystemEliteIdRequestsPort.loadAll()).thenReturn(Collections.emptyList());
+        when(loadAllSystemEliteIdRequestsPort.loadAll()).thenReturn(Collections.emptyList());
         when(loadSystemsByFilterPort.loadByFilter(any())).thenReturn(Collections.emptyList());
 
         underTest.cleanUpObsolete();
@@ -77,7 +115,7 @@ public class CleanUpObsoleteSystemEliteIdRequestsUseCaseTest {
         System system1 = mock(System.class);
         when(system1.getName()).thenReturn("Gamma");
 
-        when(allSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
+        when(loadAllSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
         when(loadSystemsByFilterPort.loadByFilter(any())).thenReturn(List.of(system1, system2, system3));
 
         underTest.cleanUpObsolete();
@@ -95,7 +133,7 @@ public class CleanUpObsoleteSystemEliteIdRequestsUseCaseTest {
         System system1 = mock(System.class);
         when(system1.getName()).thenReturn("Alpha");
 
-        when(allSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
+        when(loadAllSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
         when(loadSystemsByFilterPort.loadByFilter(any())).thenReturn(List.of(system1));
 
         underTest.cleanUpObsolete();
@@ -117,7 +155,7 @@ public class CleanUpObsoleteSystemEliteIdRequestsUseCaseTest {
         System system2 = mock(System.class);
         when(system2.getName()).thenReturn("Beta");
 
-        when(allSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
+        when(loadAllSystemEliteIdRequestsPort.loadAll()).thenReturn(List.of(request1, request2));
         when(loadSystemsByFilterPort.loadByFilter(any())).thenReturn(List.of(system1, system2));
 
         underTest.cleanUpObsolete();
